+++ ActiveRecord.php
@@ -38,6 +38,18 @@ use yii\helpers\StringHelper;
  */
 class ActiveRecord extends BaseActiveRecord
 {
+    private $_oldAttributes;
+    
+    protected $_keyPrefix;
+    protected $_PrimaryKey;
+    protected $_expireTime;
+    protected $_lonlat;
+
+    const TYPE_ID       = 's';
+    const TYPE_KEY      = 'a';
+    const TYPE_GEO      = 'g';
+    const TYPE_EXPIRE   = 'e';
+    
     /**
      * Returns the database connection used by this AR class.
      * By default, the "redis" application component is used as the database connection.
@@ -58,6 +70,19 @@ class ActiveRecord extends BaseActiveRecord
         return Yii::createObject(ActiveQuery::className(), [get_called_class()]);
     }
 
+    public function save($runValidation = true, $attributeNames = null)
+    {
+        if (!$attributeNames) {
+            $attributeNames = $this->safeAttributes();
+        }
+        return parent::save($runValidation, $attributeNames);
+    }    
+    
+    public function safeAttributes() 
+    {
+        return null;
+    }
+    
     /**
      * Returns the primary key name(s) for this AR class.
      * This method should be overridden by child classes to define the primary key.
@@ -92,9 +117,48 @@ class ActiveRecord extends BaseActiveRecord
     {
         return Inflector::camel2id(StringHelper::basename(get_called_class()), '_');
     }
-
+    
     /**
-     * @inheritdoc
+     * Return stored keyPrefix for best performance
+     * @return type
+     */
+    public function keyPrefixStored() 
+    {
+        if (!$this->_keyPrefix) {
+            $this->_keyPrefix = static::keyPrefix();
+        }
+        return $this->_keyPrefix;
+    }
+
+    protected function getPKey($prefix, &$values = [])
+    {
+        $db = static::getDb();
+        $pk = [];
+
+        foreach ($this->primaryKey() as $key) {
+            $pk[$key] = $values[$key] = $this->getAttribute($key);
+            if ($pk[$key] === null) {
+                // use auto increment if pk is null
+                $pkey = static::buildKeyPrefix($key, $prefix, static::TYPE_ID);
+                $pk[$key] = $values[$key] = $db->executeCommand('INCR', [$pkey]);
+                $this->setAttribute($key, $pk[$key]);
+            } elseif (is_numeric($pk[$key])) {
+                // if pk is numeric update auto increment value
+                $pkey = static::buildKeyPrefix($key, $prefix, static::TYPE_ID);
+                $currentPk = $db->executeCommand('GET', [$pkey]);
+                if ($pk[$key] > $currentPk) {
+                    $db->executeCommand('SET', [$pkey, $pk[$key]]);
+                }
+            }
+        }
+        return static::buildKey($pk);
+    }
+    
+    /**
+     * Insert record
+     * @param bool $runValidation
+     * @param array $attributes
+     * @return boolean
      */
     public function insert($runValidation = true, $attributes = null)
     {
@@ -104,29 +168,24 @@ class ActiveRecord extends BaseActiveRecord
         if (!$this->beforeSave(true)) {
             return false;
         }
+        $prefix = $this->keyPrefixStored();
         $db = static::getDb();
         $values = $this->getDirtyAttributes($attributes);
-        $pk = [];
-        foreach ($this->primaryKey() as $key) {
-            $pk[$key] = $values[$key] = $this->getAttribute($key);
-            if ($pk[$key] === null) {
-                // use auto increment if pk is null
-                $pk[$key] = $values[$key] = $db->executeCommand('INCR', [static::keyPrefix() . ':s:' . $key]);
-                $this->setAttribute($key, $values[$key]);
-            } elseif (is_numeric($pk[$key])) {
-                // if pk is numeric update auto increment value
-                $currentPk = $db->executeCommand('GET', [static::keyPrefix() . ':s:' . $key]);
-                if ($pk[$key] > $currentPk) {
-                    $db->executeCommand('SET', [static::keyPrefix() . ':s:' . $key, $pk[$key]]);
-                }
-            }
-        }
-        // save pk in a findall pool
-        $db->executeCommand('RPUSH', [static::keyPrefix(), static::buildKey($pk)]);
 
-        $key = static::keyPrefix() . ':a:' . static::buildKey($pk);
+        $key = $this->getPKey($prefix);
+        $time = ''; $type = null;
+        if ($this->_expireTime) {
+            $time = time()+$this->_expireTime;
+            $type = static::TYPE_EXPIRE;
+        }
+        // save unique pk in a findall pool if key isn't expire
+        $pool = static::buildKeyPrefix($time, $prefix, $type);
+        $db->executeCommand('SADD', [$pool, $key]);
+
+        $this->_PrimaryKey = static::buildKeyPrefix($key, $prefix, static::TYPE_KEY);
         // save attributes
-        $setArgs = [$key];
+        $setArgs = [$this->_PrimaryKey];
+        
         foreach ($values as $attribute => $value) {
             // only insert attributes that are not null
             if ($value !== null) {
@@ -141,6 +200,35 @@ class ActiveRecord extends BaseActiveRecord
         if (count($setArgs) > 1) {
             $db->executeCommand('HMSET', $setArgs);
         }
+        
+        if ($this->_expireTime) {
+            $args = [
+                'key' => $this->_PrimaryKey,
+                'expire' => $this->_expireTime,
+            ];
+            $db->executeCommand('EXPIRE', $args);
+            $args['key'] = $pool;
+            $db->executeCommand('EXPIRE', $args);
+        }
+        
+        if ($this->_lonlat) {
+            $k = $prefix.':'.static::TYPE_GEO;
+            if ($this->_expireTime) {
+                $k .= ':'.$time;
+            }
+            $geo = [
+                $k,
+                $this->_lonlat['lon'],
+                $this->_lonlat['lat'],
+                $key,
+            ];
+            $db->executeCommand('GEOADD', $geo);
+            if ($this->_expireTime) {
+                $args['key'] = $k;
+                $db->executeCommand('EXPIRE', $args);
+            }
+        }
+        
 
         $changedAttributes = array_fill_keys(array_keys($values), null);
         $this->setOldAttributes($values);
@@ -150,29 +238,72 @@ class ActiveRecord extends BaseActiveRecord
     }
 
     /**
-     * Updates the whole table using the provided attribute values and conditions.
-     * For example, to change the status to be 1 for all customers whose status is 2:
-     *
-     * ~~~
-     * Customer::updateAll(['status' => 1], ['id' => 2]);
-     * ~~~
-     *
-     * @param array $attributes attribute values (name-value pairs) to be saved into the table
-     * @param array $condition the conditions that will be put in the WHERE part of the UPDATE SQL.
-     * Please refer to [[ActiveQuery::where()]] on how to specify this parameter.
-     * @return integer the number of rows updated
+     * @see update()
+     * @param array $attributes attributes to update
+     * @return integer number of rows updated
+     * @throws StaleObjectException
      */
-    public static function updateAll($attributes, $condition = null)
+    protected function updateInternal($attributes = null)
     {
-        if (empty($attributes)) {
+        if (!$this->beforeSave(false)) {
+            return false;
+        }
+        $values = $this->getDirtyAttributes($attributes);
+        if (empty($values)) {
+            $this->afterSave(false, $values);
             return 0;
         }
+        $pk = $this->getOldPrimaryKey(true);
+        $lock = $this->optimisticLock();
+        if ($lock !== null) {
+            $values[$lock] = $this->$lock + 1;
+            $pk[$lock] = $this->$lock;
+        }
+
+        // We do not check the return value of updateAll() because it's possible
+        // that the UPDATE statement doesn't change anything and thus returns 0.
+        $keys = static::updatePks($values, [$pk], $this->_expireTime);
+        $rows = count($keys);
+        if ($rows) {
+            $this->_PrimaryKey = reset($keys);
+        }
+
+        if ($lock !== null && !$rows) {
+            throw new StaleObjectException('The object being updated is outdated.');
+        }
+
+        if (isset($values[$lock])) {
+            $this->$lock = $values[$lock];
+        }
+
+        $changedAttributes = [];
+        foreach ($values as $name => $value) {
+            $changedAttributes[$name] = isset($this->_oldAttributes[$name]) ? $this->_oldAttributes[$name] : null;
+            $this->_oldAttributes[$name] = $value;
+        }
+        $this->afterSave(false, $changedAttributes);
+
+        return $rows;
+    }    
+    
+    /**
+     * Update records by primary keys
+     * @param array $attributes attribute values (name-value pairs) to be saved into the table
+     * @param array $pks the primary keys
+     * @param integer|null $expire time in seconds
+     * @return integer the number of rows updated
+     */
+    protected static function updatePks($attributes, $pks, $expire = null)
+    {
+        $prefix = static::keyPrefix();
         $db = static::getDb();
-        $n = 0;
-        foreach (self::fetchPks($condition) as $pk) {
+        $n = [];
+
+        foreach ($pks as $pk) {
             $newPk = $pk;
             $pk = static::buildKey($pk);
-            $key = static::keyPrefix() . ':a:' . $pk;
+            $key = static::buildKeyPrefix($pk, $prefix, static::TYPE_KEY);
+            
             // save attributes
             $delArgs = [$key];
             $setArgs = [$key];
@@ -191,21 +322,23 @@ class ActiveRecord extends BaseActiveRecord
                 }
             }
             $newPk = static::buildKey($newPk);
-            $newKey = static::keyPrefix() . ':a:' . $newPk;
             // rename index if pk changed
             if ($newPk != $pk) {
-                $db->executeCommand('MULTI');
+                $newKey = static::buildKeyPrefix($newPk, $prefix, static::TYPE_KEY);
+                static::beginTransaction();
                 if (count($setArgs) > 1) {
                     $db->executeCommand('HMSET', $setArgs);
                 }
                 if (count($delArgs) > 1) {
                     $db->executeCommand('HDEL', $delArgs);
                 }
-                $db->executeCommand('LINSERT', [static::keyPrefix(), 'AFTER', $pk, $newPk]);
-                $db->executeCommand('LREM', [static::keyPrefix(), 0, $pk]);
+                $db->executeCommand('SADD', [$prefix, $newPk]);
+                $db->executeCommand('SREM', [$prefix, 0, $pk]);
                 $db->executeCommand('RENAME', [$key, $newKey]);
-                $db->executeCommand('EXEC');
+
+                static::commitTransaction();
             } else {
+                $newKey = $key;
                 if (count($setArgs) > 1) {
                     $db->executeCommand('HMSET', $setArgs);
                 }
@@ -213,10 +346,41 @@ class ActiveRecord extends BaseActiveRecord
                     $db->executeCommand('HDEL', $delArgs);
                 }
             }
-            $n++;
+            if ($expire) {
+                $args = [
+                    'key' => $newPk,
+                    'expire' => $expire,
+                ];
+                $db->executeCommand('EXPIRE', $args);
+            }            
+            $n[] = $newKey;
         }
 
-        return $n;
+        return $n;        
+    }
+
+        /**
+     * Updates the whole table using the provided attribute values and conditions.
+     * For example, to change the status to be 1 for all customers whose status is 2:
+     *
+     * ~~~
+     * Customer::updateAll(['status' => 1], ['id' => 2]);
+     * ~~~
+     *
+     * @param array $attributes attribute values (name-value pairs) to be saved into the table
+     * @param array $condition the conditions that will be put in the WHERE part of the UPDATE SQL.
+     * Please refer to [[ActiveQuery::where()]] on how to specify this parameter.
+     * @return integer the number of rows updated
+     */
+    public static function updateAll($attributes, $condition = null)
+    {
+        if (empty($attributes)) {
+            return 0;
+        }
+        $pks = static::fetchPks($condition);
+        $keys = static::updatePks($attributes, $pks);
+        
+        return count($keys);
     }
 
     /**
@@ -240,8 +404,9 @@ class ActiveRecord extends BaseActiveRecord
         }
         $db = static::getDb();
         $n = 0;
-        foreach (self::fetchPks($condition) as $pk) {
-            $key = static::keyPrefix() . ':a:' . static::buildKey($pk);
+        foreach (static::fetchPks($condition) as $pk) {
+            $key = static::buildKeyPrefix(static::keyPrefix(), static::buildKey($pk), static::TYPE_KEY);
+
             foreach ($counters as $attribute => $value) {
                 $db->executeCommand('HINCRBY', [$key, $attribute, $value]);
             }
@@ -267,29 +432,37 @@ class ActiveRecord extends BaseActiveRecord
      */
     public static function deleteAll($condition = null)
     {
-        $pks = self::fetchPks($condition);
+        $pks = static::fetchPks($condition);
         if (empty($pks)) {
             return 0;
         }
 
         $db = static::getDb();
         $attributeKeys = [];
-        $db->executeCommand('MULTI');
+        static::beginTransaction();
+        $prefix = static::keyPrefix();
         foreach ($pks as $pk) {
             $pk = static::buildKey($pk);
-            $db->executeCommand('LREM', [static::keyPrefix(), 0, $pk]);
-            $attributeKeys[] = static::keyPrefix() . ':a:' . $pk;
+            $db->executeCommand('SREM', [$prefix, 0, $pk]);
+            $attributeKeys[] = static::buildKeyPrefix($prefix, $pk, static::TYPE_KEY);
         }
         $db->executeCommand('DEL', $attributeKeys);
-        $result = $db->executeCommand('EXEC');
+        $result = static::commitTransaction();
 
         return end($result);
     }
 
-    private static function fetchPks($condition)
+    /**
+     * Find keys in key pool
+     * @param type $condition
+     * @return array of keys
+     */
+    protected static function fetchPks($condition = [])
     {
         $query = static::find();
-        $query->where($condition);
+        if ($condition) {
+            $query->where($condition);
+        }
         $records = $query->asArray()->all(); // TODO limit fetched columns to pk
         $primaryKey = static::primaryKey();
 
@@ -323,10 +496,11 @@ class ActiveRecord extends BaseActiveRecord
             }
             ksort($key); // ensure order is always the same
             $isNumeric = true;
-            foreach ($key as $value) {
+            foreach ($key as $k => $value) {
                 if (!is_numeric($value)) {
                     $isNumeric = false;
                 }
+                $key[$k] = strval($value);
             }
             if ($isNumeric) {
                 return implode('-', $key);
@@ -335,4 +509,117 @@ class ActiveRecord extends BaseActiveRecord
 
         return md5(json_encode($key));
     }
+    
+    /**
+     * Build key whith key prefix, for example User:a:1
+     * @param string $key
+     * @param string $prefix 
+     * @param type $type
+     * @return string 
+     */
+    public static function buildKeyPrefix($key, $prefix = '', $type = null)
+    {
+        if (!$prefix) {
+            $prefix = static::keyPrefix();
+        }
+        if (!$type) {
+            return $prefix;
+        }
+        
+        return "$prefix:$type:$key";
+    }
+    
+    /**
+     * Begin redis transaction
+     */
+    public static function beginTransaction()
+    {
+        $db = static::getDb();
+        $db->executeCommand('MULTI');
+    }
+
+    /**
+     * Commit redis transaction
+     * @return type
+     */
+    public static function commitTransaction()
+    {
+        $db = static::getDb();
+        return $db->executeCommand('EXEC');
+    }
+    
+    /**
+     * Returns the remaining time to live of a key that has a timeout.
+     * @return type
+     */
+    public function ttl()
+    {
+        $db = static::getDb();
+
+        if (!$this->_PrimaryKey) {
+            $pk = [];
+            foreach ($this->primaryKey() as $key) {
+                $pk[$key] = $this->getAttribute($key);
+            }
+            $this->_PrimaryKey = static::buildKeyPrefix(static::buildKey($pk), $this->keyPrefixStored(), static::TYPE_KEY);
+        }
+
+        return $db->executeCommand('TTL', [$this->_PrimaryKey]);
+    }
+    
+    /**
+     * Set a timeout on key
+     * @param int $expire
+     */
+    public function expire($expire)
+    {
+        $this->_expireTime = $expire;
+    }
+
+    /**
+     * Adds the specified geospatial items (latitude, longitude) to the key
+     * @param type $params
+     */
+    public function geo($params)
+    {
+        $this->_lonlat = $params;
+    }
+    
+    /**
+     * Adds the specified geospatial items (latitude, longitude, name) to the specified key
+     * @param string $key
+     * @param array $members
+     */
+    public static function geoAdd($key, $members)
+    {
+        $db = static::getDb();
+        $db->executeCommand('GEOADD', array_merge([$key], $members));
+    }
+    
+    /**
+     * Return the members of a sorted set populated with geospatial information
+     * @param strind $key
+     * @param float $lon
+     * @param float $lat
+     * @param string $radius - the radius for example '1 km'
+     * @return type
+     */
+    public static function findGeoRadius($key, $lon, $lat, $radius)
+    {
+        $db = static::getDb();
+        $params = explode(' ', $radius);
+        return $db->executeCommand('GEORADIUS', array_merge([$key, $lon, $lat], $params));
+    }
+    
+    /**
+     * Returns all keys matching pattern
+     * @param type $mask
+     * @return array of keys
+     */
+    public static function findKeys($mask)
+    {
+        $db = static::getDb();
+        return $db->executeCommand('KEYS', [$mask]);
+    }
+
 }