+++ CHANGELOG.md
@@ -1,6 +1,46 @@
 Yii Framework 2 redis extension Change Log
 ==========================================
 
+2.0.6.1 development fork
+-----------------------
+ATTENTION!!! 
+This fork use incompatible storage for key pool (SADD, SCARD, SMEMBERS)
+
+ - Bug #18 Fixed integer-string key bug
+ - Bug #20 Fixed findall pool duplicating keys
+ - Bug #33 invalid expire time?
+ - Enh #74 Added expire time
+ - Added expire and geo records
+ - optimized code
+
+ActiveRecord:
+ - Added consts TYPE_ID, TYPE_KEY, TYPE_GEO and TYPE_EXPIRE for prefixs
+ - Method save() get safeAttributes() as $attributeNames parames
+ - For best performance added _keyPrefix and _PrimaryKey properties;
+ - Overwrite BaseActiveRecord method updateInternal and added static::updatePks. Don't use self::fetchPks() for find key because updateInternal already have [pk]
+ - Change RPUSH for SADD for unique findall
+ - Added beginTransaction and commitTransaction methods
+ - Added ttl method
+ - Added expire method
+ - Added georadius method
+ - Added geoAdd method
+ - Added findGeoRadius method
+ - Added findKeys method
+
+ActiveQuery
+ - Overwrite method createModels for call $model->afterFind() without addition foreach()
+ - Union methods all() and one(), clear code
+ - Added expire, withExpiring and georadius methods
+ - Added getRows for converting redis lists to key->val
+ - for best performance added field _PrimaryKey for all records
+
+LuaScriptBuilder
+ - optimized code, clear build key for each BuildMethod
+ - Added methods geoRadius expire and allpks for build keys and get results
+ - Change all keys "key .. ':a:' .. pk" on "pkey"
+Cache
+ - remove cast to int
+
 2.0.6 under development
 -----------------------